Updated dataset storage to tag individual data variables according to which class created them.
The load() function allows individual classes to load only the data
that they create, while the test manager classes load all data into
their own ds_result and then distribute this to all measurement and
condition classes according to which class the data is created by.

This means that loading data back should exactly replicate all
the individual ds_results datasets in all object, to the same
state they were in when the data was saved.

 On branch master
 Changes to be committed:
	modified:   tmpl/tmpl_core.py
	modified:   unit_test/example_test_setup.py
	modified:   unit_test/test_example_resistor_test.py